components/bigquery-components/pyproject.toml
[project]
name = "bigquery-components"
authors = [
    {name = "Paul Owe", email = "paul.owe@owegroup.org"},
]
description = "BigQuery components"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.11"
dynamic = ["version"]
dependencies = [
    "kfp >= 2.0.0,<3.0.0",
    "pandas >= 1.3.0,<2.0.0",
]

[project.optional-dependencies]
tests = [
    "google-cloud-bigquery>=2.30.0",
    "pytest >= 7.3.1,<8.0.0",
    "pytest-cov >= 4.0.0,<5.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta:__legacy__"

[tool.flake8]
max-line-length = 88
per-file-ignores = [
    "E203",
    "F841"
]
ignore = [
    "E203",
    "F841",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = "tests"
junit_family = "xunit2"
addopts = "--cov=src --cov-report=term-missing --cov-report=html --cov-report=xml --cov-fail-under=80"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
